yolo_api/requirements.txt
numpy
matplotlib
# opencv-python, avoid import error for libGL
opencv-python-headless
pillow
flask